// lib/markdown/types.d.ts
export interface Parent extends UnistParent {
  children: Array<Content>;
}

export interface Literal extends UnistLiteral {
  value: string;
}

export interface Root extends Parent {
  type: 'root';
}
// ...
export interface ListItem extends Parent {
  type: 'listItem';
  checked?: boolean;
  spread?: boolean;
  children: BlockContent[];
}

export interface Table extends Parent {
  type: 'table';
  align?: Array<'left' | 'right' | 'center' | null>;
  children: TableContent[];
}

export interface TableRow extends Parent {
  type: 'tableRow';
  children: RowContent[];
}

export interface TableCell extends Parent {
  type: 'tableCell';
  children: PhrasingContent[];
}
// ...
export interface Link extends Parent, Resource {
  type: 'link';
  children: StaticPhrasingContent[];
}

export interface Image extends Node, Resource, Alternative {
  type: 'image';
}

export interface LinkReference extends Parent, Reference {
  type: 'linkReference';
}

export interface ImageReference extends Node, Reference, Alternative {
  type: 'imageReference';
}

export interface Footnote extends Parent {
  type: 'footnote';
  children: PhrasingContent[];
}

export interface FootnoteReference extends Node, Association {
  type: 'footnoteReference';
}

// Contents
export type BlockContent =
  | Paragraph
  | Heading
  | ThematicBreak
  | Blockquote
  | List
  | Table
  | HTML
  | Code
  | HighlightedCode;
export type FrontmatterContent = YAML;
export type DefinitionContent = Definition | FootnoteDefinition;
export type ListContent = ListItem;
export type TableContent = TableRow;
export type RowContent = TableCell;
export type StaticPhrasingContent =
  | Text
  | Emphasis
  | Strong
  | Delete
  | HTML
  | InlineCode
  | Break
  | Image
  | ImageReference
  | Footnote
  | FootnoteReference;
export type PhrasingContent = StaticPhrasingContent | Link | LinkReference;
export type TopLevelContent = BlockContent | FrontmatterContent | DefinitionContent;
export type Content = Root | TopLevelContent | ListContent | TableContent | RowContent | PhrasingContent;